requirements.txt
annotated-types==0.7.0
anyio==4.6.0
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.0
GeoAlchemy2==0.15.2
gunicorn==23.0.0
h11==0.14.0
httptools==0.6.1
idna==3.10
numpy==2.1.2
packaging==24.1
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
PyYAML==6.0.2
shapely==2.0.6
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.31.0
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
